requirements.txt
fastapi
uvicorn
python-multipart
pillow
ultralytics==8.1.0
